Improved error handling while adding note
Previous implementation addNote caught  RuntimeException and returned null which ultimately resulted in Closing NoteEditor with error message. This way if a user had entered a long data into fields, it gets lost due to an internal error and user needs to re-type everything. This is a bad way to handle error, so instead the new implementation catches error and shows an error dialog which user might dismiss and copy the entered data elsewhere.

Removed the unnecessary variable 'context'