Fix speed regression in browse screen with new schema
46c4f4d50173988cdceab8411fc4548478ababfa changed a single doProgress()
call to multiple ones. I suspect the speed improvements in that PR
came from checking cancellation at the top, and the other change was
not necessary.

Things got worse with the introduction of #11849, as the progress listener
is called it a hot loop, and it invokes colIsOpen() each time.

This change puts performance back on par with the legacy path; further
improvements should be possible in the future by switching to a recycler
view and another backend method: https://github.com/ankidroid/Anki-Android/issues/11889