Probable fix for occasional crash when editing with v3
Crash report includes the following:

Caused by: net.ankiweb.rsdroid.exceptions.BackendInvalidInputException: Invalid input: ⁨card was modified: Normal(
    New(
        NewState {
            position: 15846,
        },
    ),
) Normal(
    Learning(
        LearnState {
            remaining_steps: 2,
            scheduled_secs: 60,
        },
    ),
)⁩

This is the backend noticing that the card details have changed since
it was presented to the user. I believe what's happening here is that
UpdateNote() uses raw SQL statements to save the note, which causes the
V3 queue to be rebuilt. If a learning card became due in the mean time,
that could result in a different card being presented.

Solved by using the backend method to save a note, as it's undoable,
and doesn't clear the study queue.

Also:

- Use the backend method to change card deck, as it's undoable
- Move previously-unused setDeck() out of `decks` and into `col`, to
match desktop code.
- Refactor toBackendNote(), as it was broken and causing tests to fail
- Add a missing @BlocksSchemaUpgrade