fix: margin of Floating Action Button in Card Browser
Fixed the margin issue with FAB in Card Browser by increasing the marginEnd and marginBottom. The marginEnd initial value was 10dp and was changed to 32dp because on comparing it with the margin+padding of FAB in Decks activity.

Similarly extra 12dp was added to marginBottom when comparing with FAB in Decks activity