Enlarge size of checkbox in Tag Picker dialog (#11974)
* update `tags_item_list_dialog` to support 48x48 checkbox size

added couple of more changes like making width to `match_parent` instead of `fill_parent` as after API level 8+ it is recommended to use `match_parent`
Also changed `paddingLeft` to `paddingStart` and removed `paddingRight` as it is handled by `paddingEnd` for English language and for something like Arabic language `paddingStart` will do the same thing.

* add insets to make the drawable icons center inside the checkbox

Insets is used to make the drawable icons which are of (24x24) to be in center of the checkbox which is (48x48). The value of insets is set only for left side as by default the drawable icon inside the checkbox is attached to the left side. So making the left inset value to 12dp moves the icon by 12dp to the right which makes it to go in the center of the checkbox automatically

* update `CheckBoxTriStates.kt` to use inset drawables