Avoid runBlocking() calls in CollectionManager under Robolectric
The recommended way to test coroutines in Kotlin is to wrap tests in
runTest(). Unfortunately, doing so was tending to trigger deadlocks.

It appears to be caused by calls to runBlocking() inside runTest().
These happen fairly frequently, as legacy code fetching the collection
from CollectionHelper end up invoking a runBlocking() call inside
CollectionManager.

To work around this, I've made CollectionManager bypass the runBlocking
call, and access the collection directly.

This change caused the verifyStartupNoCollection() test to start failing,
presumably due to a change in timing. It was calling colIsOpen() and
being told the collection was open, because enableNullCollection() was
closing the collection but not preventing it from being implicitly
reopened.