Partially convert answerCard() a coroutine; report errors
Easier to follow, and exceptions when calling sched.answerCard() are
now caught, so with issues like #12179 and #12155 will result in a
message being shown to the user instead of an app crash.

The two additional runTest calls take advantage of the unconfined
dispatcher introduced in a previous commit, ensuring the coroutine
completes in time. Without adding them, the two tests fail due to a
race condition.