Migrate SearchCards to coroutines
The preloading of the topmost cards has been moved into the browser,
since it's not useful for the JS API, but this makes the
searchCardsNumberOfResultCount() test harder to port. Since the
CardCache should be removed in the future, I don't think it's worth the
effort of getting the test working again.

Closes #12223