Error handling enhancement while saving updated card.
Removed try-catch from CollectionOperations.updateCard and now any exceptions during its execution will be handled by launchCatchingTask.

Both CardBrowser and AbtractFlashCardViewer were abruptly closing in case on any exception during saving the card, which could result in loss of data and will not be a good experience if the user had edited a lot of data.

Now, instead of closing the Activity, an error dialog will be shown which user can dismiss and copy the existing edited data somewhere else.