Add individual switches to Custom sync server preferences
Before, Custom sync server preferences were a bit confusing. There was
a master switch for both of them, but you could actually disable either
of the custom servers, or both of them, by not specifying the URL.

This is an attempt to resolve the confusion by removing the master
switch, and adding individual switches to the two custom server
preferences. Also, the parent preference, that is the one that you
click to get to Custom sync server preferences, now shows the status
of both child preferences in its summary. Hopefully this is less
confusing.

Additionally, instead of showing an error dialog in the case of invalid
URL, the edit text dialog doesn't allow accepting such an URL by
disabling the Ok button. Allowing the user to okay bad input means it is
going to be discarded, and that's pretty awful. What if user's custom
sync URL is 420 characters long?

Note that the URLs are now validated using OkHttp, which is more strict.

Preference migration follows in the next commit.