Replace spinner in Add note type dialog with a single choice list
Fun fact: the original goal was to make the spinner popup window corners
more round.

The logic has not been changed, except I fixed an issue introduced in
238672fa. Before that commit the code read:

    ArrayAdapter<String> newModelAdapter = new ArrayAdapter<>(this, R.layout.dropdown_deck_item, newModelLabels);
    ...
    final boolean isStdModel = addSelectionSpinner.getSelectedItemPosition() < numStdModels;

And after it read:

    ArrayAdapter<String> newModelAdapter = new ArrayAdapter<>(this, R.layout.dropdown_deck_item, mNewModelLabels);
    ...
    final boolean isStdModel = addSelectionSpinner.getSelectedItemPosition() < mNewModelLabels.size();

As you can see, the code after always set isStdModel to true. This
commit reverts to the old behavior.