NF: makes add-on error message more specific
Remove "Invalid addon package" from each error so that it can be added only once
when printed if necessary

Ensure that if a field's value is null, it does not crash the function later.

Ensure that all name errors are listed together. This ensure that the user don't have
to correct all errors one after the other.

Ensure that all missing fields are listed. This ensure the user won't have to
guess what is missing.

Fixes: #12356