function for closing FAB when chosen `Create Deck` option
This new function is same to the already present closeFloatingActionMenu() function but the only difference is that this function does not scale the FAB due to which we don't see the scaling up/down animation on FAB which makes the FAB animation smooth.

Otherwise when we click `Create Deck` the FAB undergoes the scaling up/down animation in the background which does not looks nice.