Migrated ConfReset to Coroutines
AppCompatPreferenceActivity is not a lifecycleOwner so it doesn't has access to lifecycleScope.launch{}.  To be able to launch coroutines, we need to have a CoroutineScope(SuperviserJob + Dispatcher.Main) instance using which we can launch new children jobs. Any job started from this scope will need to be cancelled in onDestroy to avoid any memory leaks and resource abuse. Here the AppCompatPreferenceActivity extends CoroutineScope whose implementation is delegated by MainScope. We need to explicitly cancel all the jobs started from this scope, so cancel is called in onDestroy.