Change bottom navigation icon & colors
The previous commit reduced the contrast of the bottom navigation bar by
adding a semi-transparent active item indicator. The contrast wasn't
great to begin with.

With this commit, the active item indicator is changed to completely
white, and the color of the active icon is changed to the color of the
bar, effectively reversing colors. This is about the best contrast we
can get without changing the color of the toolbar itself.

The previous icons used color vs absence of it to communicate whether
they indicate front or back. With those icons, reversing colors would
be confusing as this would effectively reverse the meaning of the icon.
Therefore the icons were changed to a card-shaped rectangle with a
question mark for the front and with an exclamation mark for the back.
They names are prefixed with ic_card as I am planning to add more icons
with that prefix. Note that the icons are not auto mirrored, which is
not ideal for languages such as Arabic that use a reverse question mark.