toogleMark is now a suspend function
It is a good practice to launch coroutines from a lifecycle or android callback function.

If a function launches a coroutine using lifecycleScope it cannot be tested using standard coroutines testing library because the launched coroutine escapes the test-scope and may or may not return.

Making toogleMark suspend ensures that it can be easily tested using standard test scopes or runBlocking.