[GSoC] Coroutines migration DeleteNoteMulti (#12577)
* DeleteNoteHandler now extends TaskListenerWithContext

Copied the logic from previous parent: ListenerWithProgressBarCloseOnFalse

* DeleteNoteMulti now extends TaskDelegate not DismissNotes

DismissNotes is a wrapper that just executes the original task in db-transaction. This behaviour is now directly implemented in DeleteNoteMulti

* Extracted DeleteNoteMulti::task to deleteNoteMulti

* deleteNoteMulti: result is returned finally instead of publishing as progress

In previous implementation, result was being published using doProgress and just after function returns. So basically is is equivalent to finally returning the result.

Also notes are removed from view only if the operation results in success

* Migrated DeleteNoteMulti to Coroutines

pre and post delete behaviour has be imported from Cardbrowser.mDeleteNoteHandler.

* KotlinCleanup: CardBrowser

* Exceptions are now handled by launchCatchingTask

* Changes to UI is now done behind a progress dialog

* renamed variables to more descriptive ones

result to deleteCard
deleteNoteMulti to deleteMultipleNotes