chore: remove 'welcome dialog'
Due to Scoped Storage, we will no longer need this dialog
It was added because people were complaining about the storage
permission prompt, and this will no longer be shown

We remove this now to remove the burden on translators and a step
for developers to test a new version of the app.

Later, we'll remove this function entirely.

See: issue 5304 - scoped storage