NF: save last successful export
We save MOD and date. We intend to take it into account if sync is not activated.

MOD because if there was no change since last export, it's clear we don't risk
loosing data.

Actual time because that will ensure we can know how much time of review/card
creation can be lost.