Use fragment lifecycle in Fragment.launchCatchingTask
You might be wondering why this is here.

Well, at first in this PR there was an Activity that listened to
configuration changes. While size calculations were preserved between
orientation changes, changing night mode would require recalculation
since it recreates the activity.

So I decided to have a fragment with `setRetainInstance(true)`, which
required this change; without it, activity recreation would cancel the
calculations.

Then, since I went this far, I decided I might as well do the whole
`ViewModel` stuff. So this is not really needed for this PR. Still it's
a sensible change, so I left it here.