Migrate the ImportAdd task to a coroutine
The visibility of the withProgressDialog method from CoroutineHelpers was changed in order to use it in the coroutine for
 ImportAdd task as we need access to the progress dialog(which withProgress doesn't provide).

Showing the progress was done through a Channel which basically replaces the previous ProgressCallback class. Because
awaiting for values on a Channel suspends we need to listen for progress updates in a new coroutine.