Remove context receivers
The reason for the removal of the beautiful, amazing, and all-around
wonderful context receivers--as seen on YouTube!--is [redacted].

This also removes most of the commit Create & use a call-site withQueue.
The reason is that the alternative to using context receivers here
would be using a regular receiver:

    suspend fun Queue.deleteCollectionDirectory() { ... }

This might be about the same thing from the point of view of compiler,
but the semantics are different; a regular receiver is meant to be used
for objects that are acted upon, and Queue is not such object.
This is reflected by the IDE warning that Queue is unused.

Co-authored-by: lukstbit <52494258+lukstbit@users.noreply.github.com>