Show more stats while downloading a shared deck
Instead of just showing the progress in percentage, the download screen
now shows an estimate of the time remaining along with average download speed
in the appropriate units.

As part of this upgrade, the `download_percentage` ID has been renamed to
`download_stats`, and the placeholder text has been updated.
`mDownloadPercentageText` has been renamed to `mDownloadStatsText`.

Signed-off-by: Abhishek Kumar <abhi.kr.2100@gmail.com>