Use the new API for changing languages on the language preference
And redirect to the system language picker if Android ≥ 13

Fix issue when setting languages with regions

e.g. pt-BR, es-AR

They didn't reload correctly. This was fixed on the 1.7x version of appcompat, so this commits bumps it to its alpha version.

note: appcompat 1.7.0alpha02 needs Kotlin 1.8+ (or at least it threw a duplicate class error while I hadn't upgraded the kotlin version)