Replace notify icon file with new file suitable for dark icon
The current rotification icon remains white even if the color of status bar is white or other light one.
This commit makes the icon dark (black) in such background by replacing the
current notification star file with a new one, which is made from the
current file and is specifically designed for notification icon by Asset Studio of Android Studio.