ci: reduce PR stale time to 30 days
Previous value: 60 days

Stale PRs cause time wastage due to conflicts.
It's best to get a new developer early if the PR has value
as the PR will be fresh in the mind of reviewers.

My aim is to get the PR queue near to 0, and this change
closes or bumps a number of old PRs

As per discussion on Discord, I would like this value to be at '14'
but reducing it too quickly will overburden contributors with many
PRs open.

It's easier to reduce it incrementally

docs: https://github.com/actions/stale#all-options