build(ci): reduce PR stale timeout to 2 weeks
As agreed in Discord - we don't want the PR queue to get stale
as the number of conflicts adds to the workload

I've committed to keeping things moving and fixing up stale PRs to
a reasonable extent.

If 2 weeks turns out to be too short, I'm happy to raise it back.