Scoped Storage: Enforce Media Sync Before Migration
We need to move the user's data to /Android/ to handle scoped storage
changes. Before we perform this migration, we should ensure that users
data is safe. To do this, we enforce a sync.

A user should have completed a full media sync before the migration
occurs.

This ensures that their data is up to date and can be restored if
anything goes awry with the migration

Dialog wording as discussed on the Figma, with the exception of
changing 'sync' to 'media sync'

Issue 5304