refactor: Extract permission logic from DeckPicker
This logic will get more complex after scoped storage.
This makes it more testable

This hides complexity as permissions dialogs:
* May not be necessary (permissions already granted)
* Return permission granted/denied in an async fashion

Android provides no native way to know if permissions have been denied
permanently.

To work around this, we define a PermissionsManager which encapsulates
both concerns: `.requestDialog` returns the dialog state and accepts a
 callback if the method is async.

* Fixes a TODO where a permissions dialog is not shown

In order to use `registerForActivityResult`, registration needs to
occur on activity startup.

Uses `shouldShowRequestPermissionRationale` to determine if a
permission denial is temporary or permanent

Fixes a DeckPicker TODO on checking media