Fix remaining lint and tests isssues
Because we moved from a platform method call on a class written in Java to a method written completely in kotlin, one of the
tests failed because mockito and kotlin don't play along very well(mockito passing null values for kotlin methods although they
expect non null types). I fixed this by using the actual `intent` value used in the method call in the class for which the test
was failing(see CustomTabsHelper).