Refactor MigrationSuccessDialogFragment & add MigrationFailedDialogFragment
First of all, this moves MigrationSuccessDialogFragment to
com/ichi2/anki/ui/dialogs/ActivityAgnosticDialogs.kt along with the
related code that tracks the started activities and runs code on start
of any of them.

This also adds MigrationFailedDialogFragment. As we currently lack the
knowledge about what exactly can realistically fail and how, it just
shows the exception text and allows copying the debug info in the same
way as the notification does. It would be nice to try to account for
the most common errors here.

Also this:

  * makes the above dialogs not cancellable on touch outside;

  * adds DefaultActivityLifecycleCallbacks--same as
    Application.ActivityLifecycleCallbacks but with all methods having
    default implementations--for simplicity;

  * adds Activity.canProperlyShowSnackbars() to determine if a
    snackbar, if shown in an activity, will animate properly, or if
    a toast should be shown instead;

  * fixes a slight issue in Snackbars.kt where an activity with a view
    with the ID root_layout would be treated as one that can properly
    show a snackbar--the code wasn't checking whether or not the view
    was CoordinatorLayout so it could produce the wrong result.