Use another icon for suggesting backup if no sync
(Use an alternate design for the icon of the backup suggestion dialog box
if the user doesn't use the sync feature)

The current icon with an upward arrow in a cloud alludes uploading data to cloud (AnkiWeb).
Its design is misleading for encouraging backup to those who don't use the sync feature.

By this commit, an icon which generally means "Setting backup restore" will be shown
instead of the current icon when the user hasn't logged in to their AnkiWeb account.