Use "Finish" as Timebox negative button label
The current label of the negative button of "Timebox reached" dialog is "Close",
but it misleads as if it merely closes the dialog.
(It actually finishes the review session and shows the deck picker screen.)

This commit will replace the label with "Finish",
which is used on the counterpart dialog in Anki Desktop.
The string is reused from Anki Desktop.

Incidentally, this commit will also add the title comment
to the section of Timebox in 02-string.xml, for organization.