Refactor essential file migration; use CollectionManager
By performing the essential migration inside the CollectionManager
queue, we ensure that it has exclusive access to the collection files
for the duration of the call. This should hold even for legacy code, as
CollectionHelper.getCol() blocks until the queue is free.

I found the current code somewhat hard to follow, due in large part to
its Java origins, so I've also tried to tidy it up a fair bit. I think one
of the main reasons scoped storage has stalled so long is that people were
a bit intimidated by the code (I was at least!), and I hope that by making
it a bit simpler, people will feel more comfortable jumping in and helping
to get it over the line.

Notable changes:

- The validation and copying steps were intermixed and spread through
multiple files, which made it hard to follow. The source/dest folders are
now determined/validated separately, and then the actual copying
is done in a CollectionManager method. The chain starts in deckpicker.kt:migrate(),
as before.
- The typealiases and newtypes have been removed, as I felt they made
the code harder to follow. The newtype also served to validate the folder
was a scoped storage folder, but that's now done in our validation routine.
Instead, the variable names have been renamed to be clearer, eg
 destinationDirectory: ScopedAnkiDroidDirectory -> scopedDestinationDirectory: Directory
- -wal are now copied if found. Usually they would get
removed when the collection is closed, but if there is a stale -wal file,
SQLite won't clean it up until a write action is performed on the DB, so it's
safer to include it.
- The corruption check before and after migration has been removed. For one,
it was only checking some referential integrity issues, and not actually checking
if the SQLite DB was corrupt. We could change the code to fix this, but I don't
see the point - we're already confirming the bytes of the copied files exactly match
the bytes of the source files, so the extra checks seem superfluous. Doing a full
integrity check on large collections is also quite slow, and risks hanging the
UI if legacy code happens to try access the collection.
- Removed the CollectionHelper code that was opening another instance of the backend.
- Removed the LockedCollection machinery, as CollectionManager takes its place.
- Removed the RetriableException code, as it only existed to deal with concurrent
access that can not happen with CollectionManager.

Tentatively closes #13870. The first report looked like migration being
launched twice in parallel, and the second appears to have been caused by
StudyOptionsFragment refreshing the screen (perhaps in onResume),
which called withCol.