Move essential file migration into the service
Instead of performing essential file migration in the scope of an
activity, perform it in a service. Essential file migration state
becomes a part of a flow of `MigrationService.Progress`, which is
collected in a coroutine in deck picker. In order to simplify logic,
to remove the need to connect to the service and determine if the
service is/should be running or not, the flow was made static.

Fixes a number of related issues, including:

  * The progress dialog would disappear on activity recreation, which
    can happen when changing the theme and during other configuration
    changes.

  * Migration service would be launched right after essential file
    migration. This could happen when the activity was already minimized
    by user, and could lead to crashes because of foreground service
    launch restrictions.

  * Essential file migration errors were not accompanied by a more
    detailed message.

Also, some terminology is changed throughout the code and the strings.
Words “copying” an “moving” are used for essential and media file
migration, respectively, and essential files are called “database
files” in strings.

Note that this commit makes use of `ExperimentalTime`.

See https://github.com/ankidroid/Anki-Android/issues/13868
See https://github.com/ankidroid/Anki-Android/issues/13970