NF: Remove setTitle in AbstractFlashCardViewer
It probably should not be `NF`, but here it is. Because each time that
`setTitle` is called, `updateActionBar` is called immediately after and change
the title anyway.

In practice `setTitle` was setting a wrong title in the Reviewer, as it was
using current deck instead of the name of the current card.

Worse, while moving access to `col` to coroutine, I realized that this created
a race condition, where the title changed. That was catched by automated test,
so that's great actually.

We can also decide that actually, we wanted to keep `Preview` as title all the
time, and current behavior is a bug. In which case I suggest we move
`updateActionBar` to `Reviewer` and ensure it does not deal with title anymore.