Fix reversal of correct/incorrect closing tag
On (the back side of) card templates in the current backend,
if you input an incorrect closing tag, for example, "{{#Tags}}blah{{/Ta}}",
Its preview says "Found '{/Tags}}', but expected '{{/Ta}}' ".

This commit will fix the inappropriate reversal.
In other words, the preview will say "Found '{/Ta}}', but expected '{{/Tags}}' "
as on the desktop version and in the advanced "new backend".