Fix card browser add icon appearing dark
As per the documentation,

> By default, all drawables instances loaded from the same resource
share a common state; if you modify the state of one instance, all the
other instances will receive the same modification.

In reviewer, tint is applied to the icons of overflow menu items.
Previously, the icons were not mutated, which lead to these icons
appearing dark when used in other activities.