set switch color without colorControlActivated
Changing colorControlActivated value originally had the purpose of tinting the preferences switches

Since this caused the problem of the text selector handle being the same color of the app bar on the light theme, using a more specific color selector should be preferred.

Android's color for the text handle is `colorControlActivated`, which defaults to `colorAccent` if not specified.

So, there are two ways to fix the issue:
1. Changing the handle style completely
2. Changing the switch/other widget colors

In order to keep the handle with a familiar look and compatible with Android future changes, and to make things simpler, it's better to avoid changing the handle, so this changes the switch color manually

To use the same color between the switch and other widgets (e.g. in SliderPreference), a `widgetColorActivated` attr similar to `colorControlActivated` was created