Rename export-in-progress message in new backend
[Before] "Exporting file..."
[After]    "Preparing Export..."

In Anki Desktop, "Exporting file..." is shown after the location to save the file is selected, so it is natural.
However, in AnkiDroid, this in-progress message is shown before the location (or application) to save the file is selected, so it isn't natural for users.

This commit will replace the message in new backend with one suitable for the timing.