NF: move `discardCurrentCard` call
The goal is to simplify moving to coroutine.

This method is pointless if there is no current card (as in previewer and
CardTemplatePreviewer); so it makes sense to move it to Reviewer only.

It is also pointless if there is no more card to review, as there is no card to
discard. So it only has to be called when the user manually close the
reviewer. That is, when they press back.

Moving this call to `onBackPressed` at the very least ensure that the blocking
operation only occurs when it's needed.

More importantly, once we move to coroutine, either this discard should be able
to survive the current lifecycle, which we usually avoids in AnkiDroid, or,
simpler, we need to actually wait for the discard to occur before closing
it. And this will be simpler if the blocking operation is launched after an
action, and not in the onDestroy method.